Add a Quaternion class
This class represents rotations and can be created directly or via
an angle + axis or from euler angles. Quaternions can then be combined
via multiplication. They can then be converted back to euler angles to
be applied to objects as rotations. This will help greatly when dealing
with angular velocities and torques for the physics engine.