Only track the Ship's position from the Shield.
Otherwise when the Shield is destroyed, we throw an error that the Ship is still being referenced but should have been destroyed.

The same effect can be achieved by not memoizing the ship (`game_object.parent`) as a method (e.g. `ship`), but it's only the position we want to track in order to update the position of the Shield.