Fix ghost renderers bug
When a renderer was destroyed it would leave the model matrix in the
render pipeline's instance buffer so it would still be drawn. This meant
that objects that had been destroyed still could be seen. To fix this I've
made sure to remove the extra data from the instance buffer when an object
is destroyed.